Move all device code to devices module
