+++ src/arch/x86/interrupts/irq.rs
@@ -1,7 +1,7 @@
-use arch::x86::device::pic;
-use arch::x86::device::serial::{COM1, COM2};
 use arch::x86::interrupts::exception::ExceptionStackFrame;
 use devices::{ps2_controller_8042, ps2_keyboard};
+use devices::pic_8259 as pic;
+use devices::uart_16550 as serial;
 
 #[allow(dead_code)]
 fn trigger(irq: u8) {
@@ -35,12 +35,12 @@ pub extern "x86-interrupt" fn cascade(_stack_frame: &mut ExceptionStackFrame) {
 
 pub extern "x86-interrupt" fn com1(_stack_frame: &mut ExceptionStackFrame) {
     pic::MASTER.lock().ack();
-    let data: u8 = COM1.lock().receive();
+    let data: u8 = serial::COM1.lock().receive();
     kprint!("{}", data as char);
 }
 
 pub extern "x86-interrupt" fn com2(_stack_frame: &mut ExceptionStackFrame) {
     pic::MASTER.lock().ack();
-    let data: u8 = COM2.lock().receive();
+    let data: u8 = serial::COM2.lock().receive();
     kprint!("{}", data as char);
 }
