+++ src/devices/mod.rs
@@ -3,3 +3,4 @@ pub mod pic_8259;
 pub mod ps2_keyboard;
 pub mod ps2_controller_8042;
 pub mod uart_16550;
+pub mod vga;
